requirements.txt
# Core packages
tensorflow>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.2

# Model libraries
xgboost>=1.5.0
lightgbm>=3.3.0
keras>=2.10.0

# Optimization
optuna>=3.0.0
hyperopt>=0.2.7

# Utility packages
tqdm>=4.62.0
psutil>=5.9.0  # For memory monitoring
gputil>=1.4.0  # For GPU monitoring (optional)
joblib>=1.1.0

# Visualization
seaborn>=0.11.0
plotly>=5.3.0

# Notebook support (optional)
ipykernel>=6.0.0
jupyter>=1.0.0

# Data processing
python-dateutil>=2.8.2
statsmodels>=0.13.0

# Other utilities
PyYAML>=6.0
urllib3>=1.26.7
requests>=2.27.0

# Core libraries
scikit-learn==1.3.0
scipy>=1.7.0
psutil==5.9.5
tqdm==4.65.0
pathlib>=1.0.1

# Statistical models
statsmodels==0.14.0
pmdarima==2.0.4
mlxtend==0.23.1

# Machine learning models
catboost==1.2.2
imbalanced-learn>=0.9.0
shap>=0.41.0

# Deep learning
tensorflow-metal==1.0.0  # macOS Metal GPU acceleration
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
filelock==3.18.0
typing-extensions>=4.8.0
sympy==1.14.0
networkx==3.4.2
jinja2==3.1.6
fsspec==2025.3.2
MarkupSafe>=2.0
mpmath<1.4,>=1.1.0
h5py>=3.1.0

# Time series forecasting
prophet==1.1.4
tsfresh>=0.19.0

# Testing and code quality
pytest>=6.2.5
pytest-cov==4.1.0
black>=21.9b0
flake8>=3.9.0
mypy>=0.910
pytest-benchmark>=4.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly>=5.3.0
graphviz>=0.20.1  # For model visualization

# Utilities
colorama>=0.4.4
numba==0.60.0
isort>=5.9.3
pre-commit>=2.15.0

# Memory management
memory-profiler>=0.60.0
py-spy>=0.3.0

# Model deployment and serialization
mlflow>=2.3.0
onnx>=1.13.0
onnxruntime>=1.14.0

# Model interpretation
lime==0.2.0.1
scikit-image==0.21.0

# Version control
GitPython>=3.1.30  # For model versioning

# System resources monitoring
gputil>=1.4.0  # For GPU monitoring

# Other dependencies
beautifulsoup4>=4.12.0  # For web scraping
pytz>=2021.3
loguru>=0.6.0  # Enhanced logging